CredRandom生成処理の実装準備 [Flash ROM関連修正は中止]
・Flash ROMを拡張し、CredRandomデータをnRF52840内に保持する実装案を却下
・代わりに、Credential IDを拡張して、CTAP2クライアントに保持させる実装とするよう方針変更
・これに伴い、Credential IDのサイズに関する変数および型を修正（２５６バイトを超える可能性が高いため）
・メンテナンス性向上のため、Flash ROMレコード領域に関するコメントを強化

+++ nRF5_SDK_v15.2.0/examples/diverta/ctap2_lib/ctap2_pubkey_credential.c
@@ -83,7 +83,7 @@ static void generate_credential_source_hash()
     
     // Public Key Credential Source + rpIdHashから
     // SHA-256ハッシュ値（32バイト）を生成
-    uint8_t hash_source_size = pubkey_cred_source_size + ctap2_rpid_hash_size;
+    size_t hash_source_size = pubkey_cred_source_size + ctap2_rpid_hash_size;
     credential_source_hash_size = sizeof(credential_source_hash);
     fido_crypto_generate_sha256_hash(
         hash_source_buffer, hash_source_size, credential_source_hash, &credential_source_hash_size);
@@ -113,7 +113,7 @@ void ctap2_pubkey_credential_generate_source(CTAP_PUBKEY_CRED_PARAM_T *param, CT
     //  34: User Id（バイト配列）のサイズ
     //  35 - n: User Id（バイト配列）
     // 
-    uint8_t offset = 1;
+    size_t offset = 1;
     memset(pubkey_cred_source, 0x00, sizeof(pubkey_cred_source));
 
     // Public Key Credential Type
@@ -152,8 +152,8 @@ void ctap2_pubkey_credential_generate_source(CTAP_PUBKEY_CRED_PARAM_T *param, CT
 
     // 暗号化対象ブロックサイズを設定
     //   AES ECBの仕様上、16の倍数でなければならない
-    int block_num = offset / 16;
-    int block_sum = block_num * 16;
+    size_t block_num = offset / 16;
+    size_t block_sum = block_num * 16;
     if (offset == block_sum) {
         pubkey_cred_source_block_size = offset;
     } else {
